Added a header files for the AST.
ast.hpp contains the classes for the AST nodes. Currently only the base
class AstNode has been implemented. The class has been templated to
allow for a polymorphic visitor pattern. Each class will have to be
  for the void type since unlike other types you can return a void.
node-visitor.hpp contains specialised default implementation of the
methods for the visitor pattern on the ast nodes. It is also templated.
The method for the astNode will not be implemented in the default implementation
and will have to be implemented in the class inheriting from NodeVisitor.
@author Sameer Jagdale