Added namespace and header files to main
Added the a namespace GCompiler for the AST and visitor classes.
Hence a using GCompiler needed to be added to the main file. All phases
of the compiler except for the parser and lexer will be under the GCompiler
namespace. The parser and lexer could not be added since they generate C code.
Eventually the flex and bison grammar would be modified to generate C++ code
allowing us to put the first two phases in the same namespace.